backend/requirements.txt
anyio==3.7.0
bidict==0.22.1
black==23.3.0
click==8.1.3
h11==0.14.0
httptools==0.5.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
peewee==3.16.2
platformdirs==3.5.1
pycryptodome==3.18.0
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
PyYAML==6.0.1
sniffio==1.3.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
